requirements.txt
datapizza-ai
datapizza-ai-clients-google
python-dotenv
